Add a option for the control channel address
This adds a option -K/--control to set the address for the iperf
control channel. This is useful for example if a udp test will saturate
th wire so that the control data will not be delivered.

The option has different meanings if used on the client or the server:
* At the server it is the initial bind address to open the listening
  control socket.
* At the client it is the target address sor the control channel
  connect.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>